Changelog: 1. Adjust the UI flow on transaction add bundle to be more optimized

+++ androidApp/src/main/kotlin/org/lelestacia/posle/screen/product_list/component/BundleItem.kt
@@ -41,7 +41,7 @@ fun BundleItem(
     onLongClick: (() -> Unit)? = null
 ) {
     val totalPrice = bundle.bundleProducts.sumOf {
-        it.sellPrice.value * it.quantity.value.toBigDecimal()
+        it.sellPrice.value * it.quantity.value
     }
 
     ElevatedCard(
